changed the folder name from experiments to data-collection-scripts
