+++ data-collection-scripts/data/csv/sqlite_sequence.csv
@@ -1,0 +1,3 @@
+name,seq
+news_articles,68636
+news_ticker_sentiment,320979
