+++ data-collection-scripts/finance_data.db
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c2887665b8ec86e13a7e2217e57be6203e67d868340cc790bfcdae06b75f2e93
+size 128364544
