--- experiments/company_overview.db
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a4e5f7ee37736e5bd4bfcab3176d904f4ecfd347f68a1be636ca25bc3e20070e
-size 69632
