--- experiments/data/csv/balance_sheet.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5cdb8531d9ce15e7a767f2b2d1a62badf739ebaa77f5414303f222fa361cccf7
-size 1809474
