--- experiments/data/csv/cash_flow.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:801330dd718b4a680c6c7b15bcf536f5c03170ca8cec42f8f81e62fc98a70ff4
-size 1144653
