--- experiments/data/csv/company_intraday_data.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e28273d9c7274abc9a9a53c29628c4bebddc8b9f86f4f6cfa65b09e8e0c6c419
-size 41297073
