--- experiments/data/csv/company_overview.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4d604062b07793f6b5f8e3d70028dd4039138f67c3603e29045cf80742ab43a1
-size 50095
