--- experiments/data/csv/income_statement.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2dba778f6d5d9ded883c4ba24e2ddd911706005df0f9cbeb557bdd21f30a2849
-size 311533
