--- experiments/data/csv/news_articles.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e60cb21188cdf10e7e2855a2ea56d187cefad2412df65379119b0456e33299cc
-size 44132098
