--- experiments/data/csv/news_ticker_sentiment.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e4d4204fe3bd8ecf012277f57bcee05f0ed4bba4e3031655aa90ef3fd2505a21
-size 14779678
