--- experiments/data/csv/sqlite_sequence.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4d18e712d5b6ea42b95c4f6db939b0e88d5865871ba18885385112070e5a4581
-size 61
